requirements.txt
flask == 1.1.1
pyaml == 19.4.1
numpy == 1.17.4
opencv-python == 4.1.2.30
torch == 1.3.1

